Replace the questionable PDF icon with a CC-BY-SA icon
The Adobe PDF icon policy at http://www.adobe.com/misc/linking.html
states that the icon may only be used with PDF documents created by an
Adobe Acrobat product. The DIG PDF files are not produced in this
manner.

Therefore, replace the current icon with a GIF derived from the CC BY-SA
Dropline Etiquette GNOME icons by Silvestre Herrerra
<silvestre.herrera@gmail.com>. Downloaded from
http://art.gnome.org/download/themes/icon/1049/ICON-DroplineEtiquette.tar.bz2

To comply with the "BY" part of BY-SA, the new GIF includes a comment
that clearly states the provenance of the icon.

Signed-off-by: Dan Scott <dscott@laurentian.ca>